fix: skip tutorial auto-start when festival is not active
Tutorial steps target home-screen elements that only render during an
active festival. Opening the app before a festival starts (or after it
ends) produced an empty walkthrough.

Auto-start is now gated on getFestivalStatus === "active". Manual
invocation remains unchanged. Extracted getFestivalStatus into
@prostcounter/shared/utils so web and mobile share the same check.
